refactor : move related Authorization exception to common exception package
